Fixed review notes
